refactor: Refactored AppointmentCalendar and Calendar components to remove redundant dispatch and useEffect hooks
