+++ src/components/calendar/AppointmentCalendar.jsx
@@ -1,30 +1,15 @@
-import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useNavigate } from "react-router-dom";
 import Calendar from "./Calendar";
 import TimeSlotSelector from "./TimeSlotSelector";
 import AppointmentActions from "./AppointmentActions";
-import { setSelectedDate } from "../../features/calendarSlice";
+import { useSelector } from "react-redux";
 
 const AppointmentCalendar = () => {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
   const { selectedDate } = useSelector((state) => state.calendar);
-
-  useEffect(() => {
-    return () => {
-      dispatch(setSelectedDate(null));
-    };
-  }, [navigate]);
-
-  const handleDateSelect = (date) => {
-    dispatch(setSelectedDate(date));
-  };
 
   return (
     <div className="p-4">
       <h2 className="mb-8 text-lg font-semibold text-navy">Booking Calendar</h2>
-      <Calendar handleDateSelect={handleDateSelect} />
+      <Calendar />
       {selectedDate && (
         <>
           <TimeSlotSelector />
